+++ src/app/core/mock-data/dashboard-stats.data.ts
@@ -1,0 +1,28 @@
+export const expectedUnreportedExpStats = {
+  count: 6,
+  sum: 16748.73,
+};
+
+export const expectedIncompleteExpStats = {
+  count: 1130,
+  sum: 1148487.57555147,
+};
+
+export const expectedReportStats = {
+  draft: {
+    count: 6,
+    sum: 93165.91,
+  },
+  report: {
+    count: 45,
+    sum: 5177243929.65219,
+  },
+  approved: {
+    count: 56,
+    sum: 28758273650702.816,
+  },
+  paymentPending: {
+    count: 4,
+    sum: 501602.12,
+  },
+};
