+++ src/app/core/mock-data/stats-response.data.ts
@@ -1,37 +1,3 @@
-export const apiTransactionUnreportedStatsRes = [
-  {
-    aggregates: [
-      {
-        function_name: 'count(tx_id)',
-        function_value: 6,
-      },
-      {
-        function_name: 'sum(tx_amount)',
-        function_value: 16748.73,
-      },
-    ],
-    dimensions: [],
-    name: 'scalar_stat',
-  },
-];
-
-export const apiTransactionIncompleteStatsRes = [
-  {
-    aggregates: [
-      {
-        function_name: 'count(tx_id)',
-        function_value: 1130,
-      },
-      {
-        function_name: 'sum(tx_amount)',
-        function_value: 1148487.57555147,
-      },
-    ],
-    dimensions: [],
-    name: 'scalar_stat',
-  },
-];
-
 export const apiReportStatsRes = {
   data: [
     {
@@ -150,33 +116,4 @@ export const apiReportStatsRes = {
     },
   ],
   url: '/v2/reports/stats',
-};
-
-export const expectedUnreportedExpStats = {
-  count: 6,
-  sum: 16748.73,
-};
-
-export const expectedIncompleteExpStats = {
-  count: 1130,
-  sum: 1148487.57555147,
-};
-
-export const expectedReportStats = {
-  draft: {
-    count: 6,
-    sum: 93165.91,
-  },
-  report: {
-    count: 45,
-    sum: 5177243929.65219,
-  },
-  approved: {
-    count: 56,
-    sum: 28758273650702.816,
-  },
-  paymentPending: {
-    count: 4,
-    sum: 501602.12,
-  },
 };
