+++ src/app/core/models/dashboard-stats.model.ts
@@ -1,0 +1,11 @@
+export interface DashboardStats {
+  count: number;
+  sum: number;
+}
+
+export interface ReportDashboardStats {
+  draft: DashboardStats;
+  report: DashboardStats;
+  approved: DashboardStats;
+  paymentPending: DashboardStats;
+}
